--- internal/client/client.go
@@ -1,3 +1,0 @@
-package client
-
-const BaseURL = "https://graph.facebook.com"
